+++ projects/igniteui-angular/migrations/update-17_0_0/index.spec.ts
@@ -205,4 +205,64 @@ describe(`Update to ${version}`, () => {
         `
         );
     });
+
+    it('Should properly rename newSelection and oldSelection property to newValue and oldValue in Combo', async () => {
+        pending('set up tests for migrations through lang service');
+        appTree.create('/testSrc/appPrefix/component/test.component.ts',
+        `
+        import { IgxComboComponent, IComboSelectionChangingEventArgs } from 'igniteui-angular';
+        export class MyClass {
+            public handleSelectionChanging(e: IComboSelectionChangingEventArgs) {
+                const newSelection = e.newSelection;
+                const oldSelection = e.oldSelection;
+            }
+        }
+        `);
+
+        const tree = await schematicRunner.runSchematic(migrationName, {}, appTree);
+
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/test.component.ts')
+        ).toEqual(
+        `
+        import { IgxComboComponent, IComboSelectionChangingEventArgs } from 'igniteui-angular';
+        export class MyClass {
+            public handleSelectionChanging(e: IComboSelectionChangingEventArgs) {
+                const newSelection = e.newValue;
+                const oldSelection = e.oldValue;
+            }
+        }
+        `
+        );
+    });
+
+    it('Should properly rename newSelection and oldSelection property to newValue and oldValue SimpleCombo', async () => {
+        pending('set up tests for migrations through lang service');
+        appTree.create('/testSrc/appPrefix/component/test.component.ts',
+        `
+        import { IgxSimpleComboComponent, ISimpleComboSelectionChangingEventArgs } from 'igniteui-angular';
+        export class MyClass {
+            public handleSelectionChanging(e: ISimpleComboSelectionChangingEventArgs) {
+                const newSelection = e.newSelection;
+                const oldSelection = e.oldSelection;
+            }
+        }
+        `);
+
+        const tree = await schematicRunner.runSchematic(migrationName, {}, appTree);
+
+        expect(
+            tree.readContent('/testSrc/appPrefix/component/test.component.ts')
+        ).toEqual(
+        `
+        import { IgxComboComponent, IComboSelectionChangingEventArgs } from 'igniteui-angular';
+        export class MyClass {
+            public handleSelectionChanging(e: IComboSelectionChangingEventArgs) {
+                const newSelection = e.newValue;
+                const oldSelection = e.oldValue;
+            }
+        }
+        `
+        );
+    });
 });